fix(config): validate tensor schedules across setup capacities
Extend SIS coverage and setup envelope sizing for tensor schedules, and update tests for tensor-by-default proof layouts so the full workspace suite passes.

+++ crates/akita-prover/src/protocol/ring_switch.rs
@@ -904,7 +904,7 @@ mod tests {
         type F = Prime128OffsetA7F7;
         const D: usize = 128;
 
-        let centered = from_fn(|i| ((37 * i as i32 + 11) % 95) - 47);
+        let centered = from_fn(|i| i64::from(((37 * i as i32 + 11) % 95) - 47));
         let ring =
             CyclotomicRing::<F, D>::from_coefficients(from_fn(|i| F::from_i64(centered[i] as i64)));
